refactor: 物件検索関連ロジックをカスタムフックに切り出し(#5)
