+++ src/hooks/stay/useStaySearchForm.ts
@@ -1,0 +1,48 @@
+import { useState, useCallback } from 'react';
+
+export type Guests = {
+  adults: number;
+  children: number;
+};
+
+const useStaySearchForm = () => {
+  const [location, setLocation] = useState<string>('');
+  const [guests, setGuests] = useState<Guests>({
+    adults: 0,
+    children: 0,
+  });
+  const handleSelectLocation = useCallback((selectLocation: string) => {
+    setLocation(selectLocation);
+  }, []);
+  const handlePlusGuests = useCallback(
+    (prop: keyof Guests) => {
+      console.log('test');
+      console.log(guests);
+      setGuests((prevGuests) => {
+        const calcValue = prevGuests[prop] + 1;
+        return { ...guests, [prop]: calcValue };
+      });
+    },
+    [guests]
+  );
+  const handleMinusGuests = useCallback(
+    (prop: keyof Guests) => {
+      setGuests((prevGuests) => {
+        const calcValue =
+          prevGuests[prop] < 1 ? prevGuests[prop] : prevGuests[prop] - 1;
+        return { ...guests, [prop]: calcValue };
+      });
+    },
+    [guests]
+  );
+
+  return {
+    location,
+    guests,
+    handleSelectLocation,
+    handlePlusGuests,
+    handleMinusGuests,
+  };
+};
+
+export default useStaySearchForm;
